+++ CHANGELOG.md
@@ -21,9 +21,9 @@
     - Added F Beta classificaion metric
     - Removed F1 Score and Informedness metric
     - Removed ISRU and Gaussian activation functions
-    - ReLU is now optionally thresholded
     - Fixed SELU derivitive computation
     - Changed adaptive optimizer default decay parameters
+    - Added AdaMax neural network Optimizer
 
 - 0.0.11-beta
     - K Means now uses mini batch GD instead of SGD